+++ scripts/army/v5_army_dispatch_adapter.gd
@@ -113,6 +113,16 @@ func get_field_tactics_read_model() -> Dictionary:
 	return city.get_field_tactics_read_model() if city != null else {}
 
 
+func preview_field_supply_transport(source_point_id: StringName) -> Dictionary:
+	var city := _get_city()
+	return city.preview_field_supply_transport(source_point_id) if city != null else {}
+
+
+func begin_field_supply_transport(source_point_id: StringName) -> Dictionary:
+	var city := _get_city()
+	return city.begin_field_supply_transport(source_point_id) if city != null else {}
+
+
 func plan_field_path(source_point_id: StringName, target_point_id: StringName, preferred_world_points: Array = [], required_road_id: StringName = &"") -> Dictionary:
 	var city := _get_city()
 	return city.plan_field_path(source_point_id, target_point_id, preferred_world_points, required_road_id) if city != null else {}